# Reduce async overhead