--- a/src/BenchmarkDotNet/Engines/EngineParameters.cs
+++ b/src/BenchmarkDotNet/Engines/EngineParameters.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading.Tasks;
 using BenchmarkDotNet.Characteristics;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Running;
@@ -12,19 +13,19 @@ namespace BenchmarkDotNet.Engines
         public static readonly IResolver DefaultResolver = new CompositeResolver(BenchmarkRunnerClean.DefaultResolver, EngineResolver.Instance);
 
         public IHost Host { get; set; }
-        public Action<long> WorkloadActionNoUnroll { get; set; }
-        public Action<long> WorkloadActionUnroll { get; set; }
+        public Func<long, IClock, ValueTask<ClockSpan>> WorkloadActionNoUnroll { get; set; }
+        public Func<long, IClock, ValueTask<ClockSpan>> WorkloadActionUnroll { get; set; }
         public Action Dummy1Action { get; set; }
         public Action Dummy2Action { get; set; }
         public Action Dummy3Action { get; set; }
-        public Action<long> OverheadActionNoUnroll { get; set; }
-        public Action<long> OverheadActionUnroll { get; set; }
+        public Func<long, IClock, ValueTask<ClockSpan>> OverheadActionNoUnroll { get; set; }
+        public Func<long, IClock, ValueTask<ClockSpan>> OverheadActionUnroll { get; set; }
         public Job TargetJob { get; set; } = Job.Default;
         public long OperationsPerInvoke { get; set; } = 1;
-        public Action GlobalSetupAction { get; set; }
-        public Action GlobalCleanupAction { get; set; }
-        public Action IterationSetupAction { get; set; }
-        public Action IterationCleanupAction { get; set; }
+        public Func<ValueTask> GlobalSetupAction { get; set; }
+        public Func<ValueTask> GlobalCleanupAction { get; set; }
+        public Func<ValueTask> IterationSetupAction { get; set; }
+        public Func<ValueTask> IterationCleanupAction { get; set; }
         public bool MeasureExtraStats { get; set; }
 
         [PublicAPI] public string BenchmarkName { get;  set; }