… src/BenchmarkDotNet/Helpers/AwaitHelper.cs
… src/BenchmarkDotNet/Helpers/ManualResetValueTaskSource.cs
… src/BenchmarkDotNet/Helpers/Reflection.Emit/IlGeneratorStatementExtensions.cs
… src/BenchmarkDotNet/Running/BenchmarkCase.cs
… src/BenchmarkDotNet/Templates/BenchmarkType.txt
… src/BenchmarkDotNet/Toolchains/InProcess.Emit.Implementation/ConsumableTypeInfo.cs
… src/BenchmarkDotNet/Toolchains/InProcess.Emit.Implementation/Emitters/ConsumableConsumeEmitter.cs
… src/BenchmarkDotNet/Toolchains/InProcess.Emit.Implementation/Emitters/ConsumeEmitter.cs
… src/BenchmarkDotNet/Toolchains/InProcess.Emit.Implementation/Emitters/RunnableEmitter.cs
… src/BenchmarkDotNet/Toolchains/InProcess.Emit.Implementation/Emitters/TaskConsumeEmitter.cs
… src/BenchmarkDotNet/Toolchains/InProcess.Emit.Implementation/Runnable/RunnableConstants.cs
… src/BenchmarkDotNet/Toolchains/InProcess.Emit.Implementation/Runnable/RunnableReflectionHelpers.cs
… src/BenchmarkDotNet/Toolchains/InProcess.NoEmit/BenchmarkAction.cs
… src/BenchmarkDotNet/Toolchains/InProcess.NoEmit/BenchmarkActionFactory.cs
… src/BenchmarkDotNet/Toolchains/InProcess.NoEmit/BenchmarkActionFactory_Implementations.cs
… src/BenchmarkDotNet/Toolchains/InProcess.NoEmit/InProcessNoEmitRunner.cs
… src/BenchmarkDotNet/Toolchains/InProcess/BenchmarkAction.cs
… src/BenchmarkDotNet/Toolchains/InProcess/BenchmarkActionFactory.cs
… src/BenchmarkDotNet/Toolchains/InProcess/BenchmarkActionFactory_Implementations.cs
… src/BenchmarkDotNet/Toolchains/InProcess/InProcessRunner.cs
… src/BenchmarkDotNet/Toolchains/Roslyn/Generator.cs
… src/BenchmarkDotNet/Validators/ExecutionValidatorBase.cs
… tests/BenchmarkDotNet.IntegrationTests/CustomEngineTests.cs
… tests/BenchmarkDotNet.IntegrationTests/InProcess.EmitTests/NaiveRunnableEmitDiff.cs
… tests/BenchmarkDotNet.IntegrationTests/InProcessEmitTest.cs
… tests/BenchmarkDotNet.IntegrationTests/InProcessTest.cs
… tests/BenchmarkDotNet.Tests/Engine/EngineFactoryTests.cs
… tests/BenchmarkDotNet.Tests/Mocks/MockEngine.cs