@YegorStepanov 2022-09-19
Since the request to add is gigantic and it still not reviewed, can you break it down into smaller ones? For example, __async InProcess.GlobalSetup__ fix, thank you.