@timcassell 2022-09-19
@YegorStepanov #1941 is a smaller PR that has the actual fixes. This PR is larger because it required a refactor to reduce the async overhead.